hotfix : 스웨거 쿼리파라미터 누락된 것 설명 추가, 쿼리 파라미터 타입 수정

+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/mapper/ProductMapper.java
@@ -98,7 +98,7 @@ public class ProductMapper {
                 .capsule(capsule)
                 .member(member)
                 .content(reviewRequest.content())
-                .score(reviewRequest.score())
+                .score((double) reviewRequest.score())
                 .build();
     }
 
@@ -109,7 +109,7 @@ public class ProductMapper {
                 .coffee(coffee)
                 .member(member)
                 .content(reviewRequest.content())
-                .score(reviewRequest.score())
+                .score((double) reviewRequest.score())
                 .build();
     }
